feat(fetch): fetch from Fastly instead of GitHub
When the request is coming from a Helix-enabled Fastly service, fetch from this service, which also works as a cache around GitHub, instead of using GitHub directly, which may fail due to outages. When the provenance of the request cannot be determined, continue using GitHub directly.

BREAKING CHANGE: fixes #348, depends on adobe/helix-publish#564